+++ eblearn/provclassifier.py
@@ -6,13 +6,14 @@ from time import time
 
 import numpy as np
 from sklearn.linear_model import SGDClassifier
+from sklearn.model_selection import GridSearchCV
 from sklearn.model_selection import GroupKFold
 from sklearn.pipeline import Pipeline
-from sklearn.model_selection import GridSearchCV
 
-from eblearn import ebpostproc, evalutils
-from eblearn.ebtransformer import EbTransformer
+from eblearn import ebpostproc
 from eblearn.ebclassifier import EbClassifier
+from eblearn.ebtransformer import EbTransformer
+from utils import evalutils
 
 # based on http://scikit-learn.org/stable/auto_examples/hetero_feature_union.html#sphx-glr-auto-examples-hetero-feature-union-py
 
@@ -203,7 +204,7 @@ class ProvisionClassifier(EbClassifier):
         evalutils.print_with_threshold(probs, y_te, overrides)
 
         self.pred_status['classifer_type'] = 'provclassifier'
-        self.pred_status['pred_status'] =  evalutils.calc_pred_status(probs, y_te)
+        self.pred_status['pred_status'] =  evalutils.calc_pred_status_with_prob(probs, y_te)
         self.pred_status['pos_threshold_status'] =  evalutils.calc_pos_threshold_prob_status(probs, y_te,
                                                                                              self.pos_threshold)
         self.pred_status['threshold_status'] = evalutils.calc_threshold_prob_status(probs, y_te, self.threshold)