+++ run_train_cust_12345.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+train_custom_prov.py --docs custom_train.txt.files --work_dir "data_from_web" --model_dir "sample_data2.model" --custom_model_dir "sample_data2.custmodel"